+++ include/core/sync_control.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <stdint.h>
 #include <stdbool.h>
 
 #include "core/types.h"
@@ -16,30 +17,30 @@ typedef struct SyncControl {
 
   LoopTrack **loop_tracks;
 
-  unsigned int track_count;
+  uint8_t track_count;
 
-  unsigned int sync_length;
-  unsigned int three_quarter_length;
-  unsigned int half_length;
-  unsigned int quarter_length;
+  uint32_t sync_length;
+  uint32_t three_quarter_length;
+  uint32_t half_length;
+  uint32_t quarter_length;
 
-  unsigned int sync_count;
+  uint32_t sync_count;
 
   SyncControlState state;
 
 } SyncControl;
 
-SyncControl *sc_create(LoopTrack **loop_tracks, unsigned int track_count);
+SyncControl *sc_create(LoopTrack **loop_tracks, uint8_t track_count);
 
-SyncControlState sc_check_track_state(SyncControl *sc, LoopTrackState lts, unsigned int record_length);
+SyncControlState sc_check_track_state(SyncControl *sc, LoopTrackState lts, uint32_t record_length);
 
-SyncControlState sc_buffer_recorded(SyncControl *sc, unsigned int record_length);
+SyncControlState sc_buffer_recorded(SyncControl *sc, uint32_t record_length);
 
 SyncControlState sc_buffer_stopped(SyncControl *sc);
 
 SyncControlState sc_buffer_cleared(SyncControl *sc);
 
-SyncTimingMessage sc_keep_sync(SyncControl *sc, unsigned int count_increase);
+SyncTimingMessage sc_keep_sync(SyncControl *sc, uint32_t count_increase);
 
 bool sc_is_syncing(SyncControl *sc);
