+++ src/tests/test_loop_track.c
@@ -1,3 +1,5 @@
+#include <stdint.h>
+
 #include "tests/minunit.h"
 
 #include "core/types.h"
@@ -19,8 +21,8 @@ char *test_loop_track_create() {
 }
 
 char *test_loop_track_state_changes() {
-  int channels = 2;
-  int frame_count = 64;
+  uint8_t channels = 2;
+  uint32_t frame_count = 64;
 
   SAMPLE *input_audio = calloc(frame_count * channels, sizeof(SAMPLE));
   SAMPLE *output_audio = calloc(frame_count * channels, sizeof(SAMPLE));
@@ -51,7 +53,7 @@ char *test_loop_track_state_changes() {
   lt_handle_action(loop_track, LoopTrack_Action_Record);
   mu_assert(loop_track->state == LoopTrack_State_Concluding, "Loop Track should be in Concluding state");
 
-  int sync_offset = 16;
+  uint32_t sync_offset = 16;
   sync_message.interval = SyncControl_Interval_Whole;
   sync_message.offset = sync_offset;
   next_state = lt_handle_audio(loop_track, sync_message, input_audio, output_audio, frame_count);
@@ -59,10 +61,10 @@ char *test_loop_track_state_changes() {
 
   mu_assert(lt_is_playing(loop_track), "Loop Track should be playing");
 
-  int expected_length = ((3 * frame_count) + sync_offset) * channels;
-  int recorded_length = lt_recorded_length(loop_track);
-  int expected_playback = (frame_count - sync_offset) * channels;
-  int playback_length = lt_playback_length(loop_track);
+  uint32_t expected_length = ((3 * frame_count) + sync_offset) * channels;
+  uint32_t recorded_length = lt_recorded_length(loop_track);
+  uint32_t expected_playback = (frame_count - sync_offset) * channels;
+  uint32_t playback_length = lt_playback_length(loop_track);
   mu_assert(recorded_length == expected_length, "Should have recorded %d not %d", expected_length, recorded_length)
   mu_assert(playback_length == expected_playback, "Should have played back %d not %d", expected_playback, playback_length)
 
@@ -90,8 +92,8 @@ char *test_loop_track_state_changes() {
 }
 
 char *test_loop_track_cancel_recording() {
-  int channels = 2;
-  int frame_count = 64;
+  uint8_t channels = 2;
+  uint32_t frame_count = 64;
 
   SAMPLE *input_audio = calloc(frame_count * channels, sizeof(SAMPLE));
   SAMPLE *output_audio = calloc(frame_count * channels, sizeof(SAMPLE));
@@ -119,8 +121,8 @@ char *test_loop_track_cancel_recording() {
 
   mu_assert(!lt_is_playing(loop_track), "Loop Track should not be playing");
 
-  int expected_length = 0;
-  int recorded_length = lt_recorded_length(loop_track);
+  uint32_t expected_length = 0;
+  uint32_t recorded_length = lt_recorded_length(loop_track);
   mu_assert(recorded_length == expected_length, "Should have recorded %d not %d", expected_length, recorded_length)
 
   lt_handle_action(loop_track, LoopTrack_Action_Playback);
